add transform stuff to init & dev cmd
