+++ apps/framework-cli/src/cli/routines/initialize.rs
@@ -24,6 +24,7 @@ impl Routine for InitializeProject {
         let run_mode: RunMode = self.run_mode;
 
         CreateInternalTempDirectoryTree::new(run_mode, self.project.clone()).run(run_mode)?;
+        CreateDenoDirectory::new(self.project.clone()).run(run_mode)?;
 
         self.project.setup_app_dir().map_err(|err| {
             RoutineFailure::new(
@@ -38,6 +39,7 @@ impl Routine for InitializeProject {
         CreateModelsVolume::new(self.project.clone()).run(run_mode)?;
         CreateDockerComposeFile::new(self.project.clone()).run(run_mode)?;
         CreateBaseAppFiles::new(self.project.clone()).run(run_mode)?;
+        CreateDenoFiles::new(self.project.clone()).run(run_mode)?;
 
         Ok(RoutineSuccess::success(Message::new(
             "Created".to_string(),
@@ -344,6 +346,68 @@ impl Routine for CreateBaseAppFiles {
         Ok(RoutineSuccess::success(Message::new(
             "Created".to_string(),
             "base app files".to_string(),
+        )))
+    }
+}
+
+pub struct CreateDenoDirectory {
+    project: Arc<Project>,
+}
+
+impl CreateDenoDirectory {
+    pub fn new(project: Arc<Project>) -> Self {
+        Self { project }
+    }
+}
+
+impl Routine for CreateDenoDirectory {
+    fn run_silent(&self) -> Result<RoutineSuccess, RoutineFailure> {
+        let internal_dir = self.project.internal_dir().map_err(|err| {
+            RoutineFailure::new(
+                Message::new(
+                    "Failed".to_string(),
+                    "to create .moose directory. Check permissions or contact us`".to_string(),
+                ),
+                err,
+            )
+        })?;
+
+        let deno_dir = internal_dir.join("deno");
+        match fs::create_dir_all(deno_dir.clone()) {
+            Ok(_) => Ok(RoutineSuccess::success(Message::new(
+                "Created".to_string(),
+                "deno directory".to_string(),
+            ))),
+            Err(err) => Err(RoutineFailure::new(
+                Message::new(
+                    "Failed".to_string(),
+                    format!("to create deno directory in {}", deno_dir.display()),
+                ),
+                err,
+            )),
+        }
+    }
+}
+
+pub struct CreateDenoFiles {
+    project: Arc<Project>,
+}
+
+impl CreateDenoFiles {
+    pub fn new(project: Arc<Project>) -> Self {
+        Self { project }
+    }
+}
+
+impl Routine for CreateDenoFiles {
+    fn run_silent(&self) -> Result<RoutineSuccess, RoutineFailure> {
+        self.project.create_deno_files().map_err(|err| {
+            RoutineFailure::new(Message::new("Failed".to_string(), "".to_string()), err)
+        })?;
+
+        Ok(RoutineSuccess::success(Message::new(
+            "Created".to_string(),
+            "deno files".to_string(),
         )))
     }
 }
