+++ apps/framework-cli/src/framework.rs
@@ -5,6 +5,7 @@ pub mod languages;
 pub mod readme;
 pub mod schema;
 pub mod sdks;
+pub mod transform;
 pub mod typescript;
 pub enum Insights {
     Metric,
