+++ apps/framework-cli/src/project.rs
@@ -30,6 +30,7 @@ use crate::cli::local_webserver::LocalWebserverConfig;
 use crate::framework::languages::SupportedLanguages;
 use crate::framework::readme::BASE_README_TEMPLATE;
 use crate::framework::schema::templates::BASE_MODEL_TEMPLATE;
+use crate::framework::transform::TRANSFORM_FILE;
 use crate::infrastructure::console::ConsoleConfig;
 use crate::infrastructure::olap::clickhouse::config::ClickhouseConfig;
 use crate::infrastructure::stream::redpanda::RedpandaConfig;
@@ -167,6 +168,17 @@ impl Project {
             let to_create = app_dir.join(dir);
             std::fs::create_dir_all(to_create)?;
         }
+
+        Ok(())
+    }
+
+    pub fn create_deno_files(&self) -> Result<(), std::io::Error> {
+        let deno_dir = self.internal_dir()?.join("deno");
+        let transform_file_path = deno_dir.join("transform.ts");
+
+        let mut transform_file = std::fs::File::create(transform_file_path)?;
+
+        transform_file.write_all(TRANSFORM_FILE.as_bytes())?;
 
         Ok(())
     }
